Better fix that doesn’t delay map start
As @mb12 points out, it isn’t necessary to delay the map start just to prevent this crash; `Map` should be able to deal with having no access token or JSON even while it’s running. Most of `Map` accounts for this situation, but `reloadStyle()` incorrectly assumes that one or the other is set.

This change reverts much of the refactoring in 41c92a6ed03b76e24b565db22fd2069313bc6243 but leaves in place some new accessors. It also corrects the assumption in `reloadStyle()`.