Removed redundant initializers and setters
This change removes most of the ways you used to be able to apply a style to the map. Building on #1163, `styleURL` (HTTP(S), mapbox:, asset:) is the canonical way to apply a style, and `styleID` is a convenient shorthand for Mapbox-hosted styles. Internally, we now construct asset: URLs in lieu of “bundled style names”.